Make pdump, and testutil compile
log and flate have moved out of the Rust disto.  Lots of minor changes
because of stabilization.

+++ src/libpool/pdump.rs
@@ -8,8 +8,8 @@ pub trait HexDump {
 struct Dumper {
     hex: String,
     ascii: String,
-    count: uint,
-    total_count: uint
+    count: usize,
+    total_count: usize
 }
 
 impl Dumper {
